+++ Changelog.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.9] - 2025-01-13
+
+### Summary
+
+- On focus build up the workflow engine which work close with serialize method in both frontend and backend.
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+- Fix the bug in sqlmanager.js when `DB` key value undefined cause the update statement return undefined value embed to the statement. Completed on 2025-01-13
+
+### Security
+
+[1.1.9]: https://github.com/wkloh76/oricommjs/releases/tag/1.1.9
+
 ## [1.1.8] - 2024-10-16
 
 ### Summary